refactored file management to allow to add files to projects and companies.

+++ delete.php
@@ -5,17 +5,20 @@ include 'controller.php';
 
 require_login('files');
 
-$filename = param('file');
+$rel_file = param('file');
 
-if (!$filename) error('No filename passed to delete!');
-
-$file = get_absolute_path($filename);
-
-if ($file){
-	if (param('confirm') == 'yes'){
-		delete_file($file);
-		redirect('index?path='.dirname($filename));
+if (!$rel_file) error('No filename passed to delete!');
+if (access_granted($rel_file)){
+	$file = base_dir().DS.$rel_file;
+	
+	if ($file){
+		if (param('confirm') == 'yes'){
+			delete_file($file);
+			redirect('index?path='.dirname($rel_file));
+		}
 	}
+} else {
+	error('You are not allowed to access "?"',$rel_file);
 }
 
 include '../common_templates/head.php';
@@ -23,6 +26,6 @@ include '../common_templates/main_menu.php';
 include 'menu.php';
 include '../common_templates/messages.php';
 
-if ($file) echo dialog(t('Are you sure, you want to delete "?"?',$filename),array(t('YES')=>'?file='.urlencode($filename).'&confirm=yes',t('NO')=>'index'));
+if ($file) echo dialog(t('Are you sure, you want to delete "?"?',$rel_file),array(t('YES')=>'?file='.urlencode($rel_file).'&confirm=yes',t('NO')=>'index'));
 
 include '../common_templates/closure.php'; ?>
